refactor: move notification and rename filterArray file
